feat: Add API Update Order Status  dan Payment Status

See merge request mirzapurnandi/dcafe-backend!15

+++ app/Services/Sultan/TableInfoService.php
@@ -86,21 +86,55 @@ class TableInfoService
         ];
     }
 
+    private function getLastMaxNumber($cafe_id) {
+        $result = Models\TableInfo::select('number')->where('cafe_id', $cafe_id)->latest()->get();
+        if(count($result) == 0) {
+            $max = null;
+        } else {
+            $max = $result[0]->number;
+
+            foreach ($result as $key => $value) {
+                if($value->number > $max) {
+                    $max = $value->number;
+                }
+            }
+        }
+
+        return $max;
+    }
+
     public function insertData($datas = [])
     {
         $status = false;
         $code = 200;
         $result = null;
+
         DB::beginTransaction();
         try {
-            $uuid = Str::uuid()->getHex()->toString();
+            $max_number = $this->getLastMaxNumber($datas['cafe_id']);
+            if($max_number == null) {
+                $max_number = 1;
+            } else {
+                $max_number++;
+            }
 
-            $table_info = new Models\TableInfo();
-            $table_info->cafe_id = $datas['cafe_id'];
-            $table_info->name  = $datas['name'];
-            $table_info->status = $datas['status'];
-            $table_info->uuid = $uuid;
-            $table_info->save();
+            $data = [];
+            
+            $total = ($datas['total']-1) + $max_number;
+            for ($i = $max_number; $i <= $total; $i++) {
+                $uuid = Str::uuid()->getHex()->toString();
+                $data[] = [
+                    'cafe_id' => $datas['cafe_id'],
+                    'name' => 'Meja',
+                    'number' => $i,
+                    'status' => true,
+                    'uuid' => $uuid,
+                    'created_at' => now(),
+                    'updated_at' => now(),
+                ];
+            }
+
+            $table_info = Models\TableInfo::insert($data);
 
             $result = $table_info;
             $message = "Successfully insert Table Info";
@@ -169,11 +203,19 @@ class TableInfoService
         $message = '';
         try {
             $table_info = Models\TableInfo::where('uuid', $uuid)->first();
+            
             if ($table_info) {
-                $table_info->delete();
-                $status = true;
-                $result = true;
-                $message = 'Successfully delete Table Info';
+                $last_number = $this->getLastMaxNumber($table_info->cafe_id);
+                if($last_number == $table_info->number) {
+                    $table_info->delete();
+
+                    $status = true;
+                    $result = true;
+                    $message = 'Successfully delete Table Info';
+                } else {
+                    $code = 404;
+                    $message = 'Data tidak boleh dihapus';
+                }
             } else {
                 $code = 404;
                 $message = 'Data tidak ditemukan';
